requirements.txt
earthengine-api
python-dotenv
pandas
numpy
tensorflow
scikit-learn
matplotlib
fastapi
geojson
requests
pydantic
